Fixed: run npm install/update in workspace folder

+++ app/index.js
@@ -283,11 +283,11 @@ var freekerneljsGenerator = yeoman.generators.Base.extend({
     
     configuration: function () {
         // Workspace
+        if (fs.existsSync('package.json')) fs.unlinkSync('package.json');
         this.template(template_name + '/_package.json', 'package.json');
         this.template(template_name + '/_bower.json', 'bower.json');
         this.template('_bowerrc', '.bowerrc');
-        if (fs.existsSync('Gruntfile.js'))
-            fs.unlinkSync('Gruntfile.js');
+        if (fs.existsSync('Gruntfile.js')) fs.unlinkSync('Gruntfile.js');
         this.copy('Gruntfile.js', 'Gruntfile.js');
         
         // Project folder
